+++ src/deephaven/plot/express/plots/__init__.py
@@ -6,5 +6,5 @@ from .distribution import histogram, violin, strip, box, _ecdf
 from .financial import candlestick, ohlc
 from .hierarchial import treemap, icicle, sunburst, funnel, funnel_area
 from .pie import pie
-from ._private_utils import layer
+from ._layer import layer
 from .subplots import make_subplots
