+++ src/deephaven/plot/express/plots/area.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
-from functools import partial
-from time import sleep
-
 from plotly import express as px
 
 from deephaven.table import Table
 
-from ._private_utils import default_callback, process_args, layer, partition_generator, process_partitions
-from ..deephaven_figure import generate_figure, DeephavenFigure
+from ._private_utils import process_args
+from ._update_wrapper import default_callback
+from ..deephaven_figure import DeephavenFigure
 
 
 def area(
@@ -31,6 +29,7 @@ def area(
         symbol_sequence: list[str] = None,
         symbol_map: dict[str, str] = None,
         size_sequence: list[int] = None,
+        size_map: dict[str, str] = None,
         xaxis_sequence: list[str] = None,
         yaxis_sequence: list[str] = None,
         markers: bool = False,
@@ -147,14 +146,15 @@ def area(
     args = locals()
 
     # partitioned args must be first so colors are properly reassigned
-    partitioned = process_partitions(args)
+    #partitioned = process_partitions(args)
 
-    partitioned_args = partition_generator(args, partitioned)
+    #partitioned_args = partition_generator(args, partitioned)
 
-    update_wrapper = process_args(args, {"area", "line"})
+    #update_wrapper = process_args(args, {"area", "line"})
 
-    draw_figure = partial(generate_figure, draw=px.area)
-
+    #draw_figure = partial(generate_figure, draw=px.area)
+    return process_args(args, {"area", "line", "stackable"}, px_func = px.area)
+"""
     trace_generator = None
     figs = []
     for arg in partitioned_args:
@@ -162,5 +162,4 @@ def area(
         if not trace_generator:
             trace_generator = fig.trace_generator
         figs.append(fig)
-
-    return update_wrapper(layer(*figs))
+"""
