+++ src/deephaven/plot/express/plots/financial.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from deephaven.table import Table
 
-from ._private_utils import default_callback, validate_common_args, process_args
+from ._private_utils import validate_common_args, process_args
+from ._update_wrapper import default_callback
 from ..deephaven_figure import generate_figure, draw_ohlc, draw_candlestick, DeephavenFigure
