+++ src/deephaven/plot/express/plots/scatter.py
@@ -5,7 +5,8 @@ from plotly import express as px
 from deephaven.table import Table
 
 from .distribution import attach_marginals, get_marg_args
-from ._private_utils import default_callback, process_args
+from ._private_utils import process_args
+from ._update_wrapper import default_callback
 from ..deephaven_figure import generate_figure, DeephavenFigure
